refactor: split common and platform-specific code
References: Eppie-io/Eppie-App#347
