+++ source/Eppie.App/Eppie.App.Shared/App.xaml.cs
@@ -1,11 +1,18 @@
+using System;
+using System.IO;
+using System.Threading.Tasks;
+
 using Tuvi.App.Shared.Models;
 using Tuvi.App.Shared.Services;
 using Tuvi.App.ViewModels.Services;
 using Tuvi.Core;
 using Tuvi.OAuth2;
+using Windows.Globalization;
+using Windows.Storage;
 
 #if WINDOWS_UWP
 using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
 #else 
 using Microsoft.UI.Xaml;
 #endif
@@ -17,7 +24,9 @@ namespace Eppie.App.Shared
     /// </summary>
     public partial class App : Application
     {
-        private static string DataBaseFileName = "TuviMail.db";
+        private static readonly string DataBaseFileName = "TuviMail.db";
+
+        private static readonly string DataFolder = ApplicationData.Current.LocalFolder.Path;
 
         public INavigationService NavigationService { get; private set; }
         public ITuviMail Core { get; private set; }
@@ -26,5 +35,122 @@ namespace Eppie.App.Shared
         private NotificationManager _notificationManager { get; set; }
 
         private ErrorHandler _errorHandler;
+
+
+        /// <summary>
+        /// Initializes the singleton application object.  This is the first line of authored code
+        /// executed, and as such is the logical equivalent of main() or WinMain().
+        /// </summary>
+        public App()
+        {
+            InitializeLogging();
+            InitializeComponent();
+            SubscribeToEvents();
+            ConfigureServices();
+        }
+
+        private static void InitializeLogging()
+        {
+            // ToDo: Add logging
+        }
+
+        private void SubscribeToEvents()
+        {
+            SubscribeToPlatformSpecificEvents();
+
+            UnhandledException += CurrentApp_UnhandledException;
+            TaskScheduler.UnobservedTaskException += TaskScheduler_UnobservedTaskException;
+        }
+
+        private void ConfigureServices()
+        {
+            CreateAuth();
+            CreateCore();
+
+            LocalSettingsService = new LocalSettingsService();
+            ApplicationLanguages.PrimaryLanguageOverride = LocalSettingsService.Language;
+        }
+
+        private void CreateAuth()
+        {
+            AuthProvider = AuthorizationFactory.GetAuthorizationProvider(Tuvi.App.Shared.Authorization.AuthConfig.GetAuthorizationConfiguration());
+        }
+
+        private void CreateCore()
+        {
+            var tokenRefresher = AuthorizationFactory.GetTokenRefresher(AuthProvider);
+            Core = ComponentBuilder.Components.CreateTuviMailCore(Path.Combine(DataFolder, DataBaseFileName), new Tuvi.Core.ImplementationDetailsProvider("Tuvi seed", "Tuvi.Package", "backup@system.service.tuvi.com"), tokenRefresher);
+            _notificationManager = new NotificationManager(Core, OnError);
+            Core.WipeAllDataNeeded += OnWipeAllDataNeeded;
+        }
+
+        private void DisposeCore()
+        {
+            Core.WipeAllDataNeeded -= OnWipeAllDataNeeded;
+            if (Core is IDisposable d)
+            {
+                d.Dispose();
+            }
+        }
+
+        private async void OnWipeAllDataNeeded(object sender, EventArgs e)
+        {
+            try
+            {
+                // Dispose should be before await method 
+                DisposeCore();
+                CreateCore();
+                await RemoveTempFilesAsync().ConfigureAwait(true);
+            }
+            catch (Exception ex)
+            {
+                OnError(ex);
+            }
+        }
+
+        private async Task RemoveTempFilesAsync()
+        {
+            var tempFolder = ApplicationData.Current.TemporaryFolder;
+            foreach (var item in await tempFolder.GetItemsAsync())
+            {
+                try
+                {
+                    await item.DeleteAsync(StorageDeleteOption.PermanentDelete);
+                }
+                catch (Exception e)
+                {
+                    OnError(e);
+                }
+            }
+        }
+
+#if WINDOWS_UWP
+        private void CurrentApp_UnhandledException(object sender, Windows.UI.Xaml.UnhandledExceptionEventArgs e)
+#else
+        private void CurrentApp_UnhandledException(object sender, Microsoft.UI.Xaml.UnhandledExceptionEventArgs e)
+#endif
+        {
+            try
+            {
+                e.Handled = true;
+                OnError(e.Exception);
+            }
+            catch { }
+        }
+
+        private void TaskScheduler_UnobservedTaskException(object sender, UnobservedTaskExceptionEventArgs e)
+        {
+            try
+            {
+                e.SetObserved();
+                OnError(e.Exception);
+            }
+            catch { }
+        }
+
+        private void OnError(Exception exception)
+        {
+            _errorHandler?.OnError(exception, false);
+        }
     }
 }
