+++ source/Eppie.App/Eppie.App.Shared/Eppie.App.Shared.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>Eppie.App.Shared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Authorization\AuthConfig.Uno.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Authorization\AuthConfig.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Authorization\AuthConfig.UWP.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Authorization\WindowsAuthenticationBroker.UWP.cs" />
